# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Examination & OSM Platform",
    "description": "Backend API for AI-assisted Digital Examination & On-Screen Marking Platform",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log in and obtain JWT access token (A1)",
        "description": "Authenticate user using email and password.\nReturns access_token, token_type, expires_in, and user profile.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current user profile (A2)",
        "description": "Return currently authenticated user's profile.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/answers": {
      "get": {
        "tags": [
          "Answers"
        ],
        "summary": "Get Answers List",
        "description": "N1: Filterable list of answers.\nExaminers only see answers assigned to them.",
        "operationId": "get_answers_list_api_v1_answers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Target exam ID",
              "title": "Exam Id"
            },
            "description": "Target exam ID"
          },
          {
            "name": "sheet_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by answer sheet ID",
              "title": "Sheet Id"
            },
            "description": "Filter by answer sheet ID"
          },
          {
            "name": "question_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by question ID",
              "title": "Question Id"
            },
            "description": "Filter by question ID"
          },
          {
            "name": "marking_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by marking status",
              "title": "Marking Status"
            },
            "description": "Filter by marking status"
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by assigned examiner ('me' or user ID)",
              "title": "Assigned To"
            },
            "description": "Filter by assigned examiner ('me' or user ID)"
          },
          {
            "name": "ocr_review_required",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by OCR review flag",
              "title": "Ocr Review Required"
            },
            "description": "Filter by OCR review flag"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/answers/{id}": {
      "get": {
        "tags": [
          "Answers"
        ],
        "summary": "Get Single Answer Workspace",
        "description": "N2: Full workspace payload for an answer.\nExaminers only access answers assigned to them. First visit tracks opened_at.",
        "operationId": "get_single_answer_workspace_api_v1_answers__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exams/{id}/anomalies/detect": {
      "post": {
        "tags": [
          "Anomalies"
        ],
        "summary": "Detect Anomalies",
        "description": "AN1: Run anomaly detectors for the given exam.\nIdempotent via dedupe_key.",
        "operationId": "detect_anomalies_api_v1_exams__id__anomalies_detect_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnomalyDetectRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyDetectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{id}/anomalies": {
      "get": {
        "tags": [
          "Anomalies"
        ],
        "summary": "Get Anomalies",
        "description": "AN2: List anomalies for an exam.",
        "operationId": "get_anomalies_api_v1_exams__id__anomalies_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by anomaly_type",
              "title": "Type"
            },
            "description": "Filter by anomaly_type"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity",
              "title": "Severity"
            },
            "description": "Filter by severity"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "examiner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by examiner",
              "title": "Examiner Id"
            },
            "description": "Filter by examiner"
          },
          {
            "name": "question_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by question",
              "title": "Question Id"
            },
            "description": "Filter by question"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/{id}": {
      "patch": {
        "tags": [
          "Anomalies"
        ],
        "summary": "Patch Anomaly",
        "description": "AN3: Dismiss or reopen an anomaly with a note.",
        "operationId": "patch_anomaly_api_v1_anomalies__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnomalyPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnomalyDetectRequest": {
        "properties": {
          "detectors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detectors"
          }
        },
        "type": "object",
        "title": "AnomalyDetectRequest"
      },
      "AnomalyDetectResponse": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Type"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "by_type"
        ],
        "title": "AnomalyDetectResponse"
      },
      "AnomalyListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnomalyResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "AnomalyListResponse"
      },
      "AnomalyPatchRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AnomalyPatchRequest"
      },
      "AnomalyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exam_id": {
            "type": "integer",
            "title": "Exam Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "answer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Id"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "examiner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Examiner Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "detected_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exam_id",
          "type",
          "severity",
          "answer_id",
          "question_id",
          "examiner_id",
          "score",
          "details",
          "status",
          "note",
          "detected_at"
        ],
        "title": "AnomalyResponse"
      },
      "AnswerDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exam_id": {
            "type": "integer",
            "title": "Exam Id"
          },
          "anon_code": {
            "type": "string",
            "title": "Anon Code"
          },
          "question": {
            "$ref": "#/components/schemas/AnswerQuestionInfo"
          },
          "rubric": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/AnswerPageInfo"
            },
            "type": "array",
            "title": "Pages"
          },
          "is_attempted": {
            "type": "boolean",
            "title": "Is Attempted"
          },
          "ocr": {
            "$ref": "#/components/schemas/AnswerOCRResponse"
          },
          "ai": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ai"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "marking_status": {
            "type": "string",
            "title": "Marking Status"
          },
          "final_marks": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Marks"
          },
          "final_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Source"
          },
          "anomalies": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anomalies"
          },
          "moderation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation"
          },
          "navigation": {
            "$ref": "#/components/schemas/AnswerNavigationInfo"
          },
          "student": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exam_id",
          "anon_code",
          "question",
          "pages",
          "is_attempted",
          "ocr",
          "ai",
          "marking_status",
          "navigation"
        ],
        "title": "AnswerDetailResponse"
      },
      "AnswerListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnswerRowResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "AnswerListResponse"
      },
...
      "AnswerRowResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "anon_code": {
            "type": "string",
            "title": "Anon Code"
          },
          "question_number": {
            "type": "string",
            "title": "Question Number"
          },
          "marking_status": {
            "type": "string",
            "title": "Marking Status"
          },
          "ocr_status": {
            "type": "string",
            "title": "Ocr Status"
          },
          "ai_status": {
            "type": "string",
            "title": "Ai Status"
          },
          "ocr_review_required": {
            "type": "boolean",
            "title": "Ocr Review Required"
          },
          "final_marks": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Marks"
          },
          "max_marks": {
            "type": "number",
            "title": "Max Marks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "anon_code",
          "question_number",
          "marking_status",
          "ocr_status",
          "ai_status",
          "ocr_review_required",
          "max_marks"
        ],
        "title": "AnswerRowResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "LoginResponse"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "examiner",
              "moderator"
            ],
            "title": "Role"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "role",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}